backend/requirements.txt
fastapi

pydantic

websockets
transformers
datasets[audio]
langchain_mongodb
pillow
accelerate
aiohttp
datasets
requests
uvicorn
pyttsx3
openai
faster-whisper
python-multipart
langchain_experimental
langchain
pygame
google-search-results
langchain_openai
sentence-transformers
moviepy
aiofiles
open_interpreter
noisereduce
pydub
asyncio
